refactor: move 'get_model_from_config'
  ... from 'agents' to 'config.agents', breaking a cycle and
  removing a 'layer' violation:

  'config.skills._SkillConfigModelBase.model_or_name' was calling it
  in the 'agents' module.
