+++ src/soliplex/config/skills.py
@@ -15,7 +15,6 @@ from haiku.skills import discovery as hs_discovery
 from haiku.skills import models as hs_models
 from pydantic_ai import models as ai_models
 
-from soliplex import agents
 from soliplex.agui import features as agui_features_module  # noqa F401
 from soliplex.config import agui as config_agui
 from soliplex.skills import bwrap_sandbox as sk_bwrap_sandbox
@@ -101,7 +100,7 @@ class _SkillConfigModelBase:
     @property
     def model_or_name(self) -> ai_models.Model | str | None:
         if self.agent_config is not None:
-            return agents.get_model_from_config(
+            return config_agents.get_model_from_config(
                 agent_config=self.agent_config,
             )
         if self.model_name is not None:
